Actualiza server node
